# src/OpencodeOai/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "opencode-oai",
    "version": "0.1.0",
    "description": "OpenAI-compatible bridge in front of a local OpenCode server. Scoped for inline completion / lightweight chat. Tool calling and workspace context are intentionally out of scope."
  },
  "servers": [
    { "url": "http://localhost:5000", "description": "Local bridge" }
  ],
  "security": [
    { "BearerAuth": [] }
  ],
  "paths": {
    "/health": {
      "get": {
        "summary": "Bridge and upstream health",
        "security": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Health" }
              }
            }
          }
        }
      }
    },
    "/v1/models": {
      "get": {
        "summary": "List available models",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ModelList" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/v1/chat/completions": {
      "post": {
        "summary": "Create a chat completion",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": { "type": "string" },
            "description": "If present, replays or coalesces same-key requests within the TTL."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChatCompletionRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Buffered response, or SSE stream when `stream=true`.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ChatCompletionResponse" }
              },
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "description": "SSE frames of ChatChunk followed by `data: [DONE]`."
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "502": {
            "description": "Upstream failure",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token; must match the `OPENCODE_PROXY_API_KEY` env var. Authentication is disabled when the env var is empty."
      }
    },
    "schemas": {
      "Health": {
        "type": "object",
        "properties": {
          "status": { "type": "string" },
          "bridge_version": { "type": "string" },
          "provider": { "type": "string" },
          "opencode": {
            "type": "object",
            "properties": {
              "connected": { "type": "boolean" },
              "version": { "type": "string" },
              "error": { "type": "string" }
            }
          }
        }
      },
      "ModelList": {
        "type": "object",
        "properties": {
          "object": { "type": "string", "const": "list" },
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Model" }
          }
        }
      },
      "Model": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "github-copilot/gpt-4o" },
          "object": { "type": "string", "const": "model" },
          "owned_by": { "type": "string" },
          "created": { "type": "integer", "format": "int64" }
        }
      },
      "ChatCompletionRequest": {
        "type": "object",
        "required": ["messages"],
        "properties": {
          "model": {
            "type": "string",
            "description": "Optional `provider/model` or bare model (uses default provider). To enable reasoning, pick a reasoning-enabled model id (e.g. `github-copilot/gpt-5-thinking`, `openai/o1`).",
            "example": "github-copilot/gpt-4o"
          },
          "messages": {
            "type": "array",
            "minItems": 1,
            "items": { "$ref": "#/components/schemas/ChatMessage" }
          },
          "stream": { "type": "boolean", "default": false },
          "reasoning_effort": {
            "type": "string",
            "description": "Accepted for OpenAI-client compatibility but not forwarded — OpenCode has no per-request reasoning-effort knob. Select a reasoning-enabled model via `model` instead.",
            "enum": ["low", "medium", "high"]
          }
        }
      },
      "ChatMessage": {
        "type": "object",
        "required": ["role"],
        "properties": {
          "role": {
            "type": "string",
            "enum": ["system", "user", "assistant", "tool"],
            "description": "`tool` messages are silently dropped."
          },
          "content": {
            "oneOf": [
              { "type": "string" },
              {
                "type": "array",
                "items": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "type": { "type": "string", "const": "text" },
                        "text": { "type": "string" }
                      }
                    },
                    {
                      "type": "object",
                      "properties": {
                        "type": { "type": "string", "const": "image_url" },
                        "image_url": {
                          "type": "object",
                          "properties": {
                            "url": {
                              "type": "string",
                              "description": "Data URI (`data:image/...;base64,...`) or remote URL."
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            ]
          }
        }
      },
      "ChatCompletionResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "object": { "type": "string", "const": "chat.completion" },
          "created": { "type": "integer", "format": "int64" },
          "model": { "type": "string" },
          "choices": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "index": { "type": "integer" },
                "message": {
                  "type": "object",
                  "properties": {
                    "role": { "type": "string", "const": "assistant" },
                    "content": { "type": "string" },
                    "reasoning_content": {
                      "type": ["string", "null"],
                      "description": "Reasoning / thinking output from reasoning-enabled models. De-facto OpenAI-compat extension (DeepSeek / OpenRouter / LiteLLM shape) recognised by Continue, Cursor, and similar clients. Null when the model didn't produce reasoning."
                    }
                  }
                },
                "finish_reason": { "type": "string", "const": "stop" }
              }
            }
          },
          "usage": { "$ref": "#/components/schemas/Usage" }
        }
      },
      "Usage": {
        "type": "object",
        "properties": {
          "prompt_tokens": { "type": "integer" },
          "completion_tokens": { "type": "integer" },
          "total_tokens": { "type": "integer" }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": { "type": "string" },
              "type": { "type": "string" }
            }
          }
        }
      }
    }
  }
}